refactor(org unit tree): simplify org data hooks
