--- components/organisation-unit-tree/src/from-entries.js
@@ -1,8 +1,0 @@
-export const fromEntries = entries =>
-    entries.reduce(
-        (collection, [key, name]) => ({
-            ...collection,
-            [key]: name,
-        }),
-        {}
-    )
