+++ components/organisation-unit-tree/src/helpers/index.js
@@ -1,0 +1,1 @@
+export { sortNodeChildrenAlphabetically } from './sort-node-children-alphabetically.js'
