+++ components/organisation-unit-tree/src/helpers/sort-node-children-alphabetically.js
@@ -1,0 +1,14 @@
+export const sortNodeChildrenAlphabetically = node => {
+    if (!node.children) return node
+    const sortedChildren = [...node.children]
+
+    sortedChildren.sort((left, right) => {
+        if (left.displayName === right.displayName) return 0
+        return left.displayName > right.displayName ? 1 : -1
+    })
+
+    return {
+        ...node,
+        children: sortedChildren,
+    }
+}
