+++ components/organisation-unit-tree/src/label/index.js
@@ -1,0 +1,1 @@
+export { Label } from './label.js'
