+++ components/organisation-unit-tree/src/use-org-data/index.js
@@ -1,0 +1,1 @@
+export { useOrgData } from './use-org-data.js'
