+++ components/organisation-unit-tree/src/use-root-org-data/index.js
@@ -1,0 +1,1 @@
+export { useRootOrgData } from './use-root-org-data.js'
